Split integration and unit tests.  Run integration tests against qz-tray.jar
